+++ packages/terra-button/src/terra-dev-site/doc/button/Button.1.doc.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/README.md';
+
 import { name, version } from '../../../../package.json';
 
-// Component Source
-import ButtonSrc from '!raw-loader!../../../../src/Button';
-
 // Example Files
+import ButtonPropsTable from '!terra-props-table-loader!../../../../src/Button';
+import ButtonSrc from '!raw-loader!../../../../src/Button';
 import ButtonBlock from '../example/ButtonBlock';
 import ButtonBlockSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/ButtonBlock';
 import ButtonAnchor from '../example/ButtonAnchor';
@@ -69,6 +69,7 @@ const DocPage = () => (
       {
         componentName: 'Button',
         componentSrc: ButtonSrc,
+        componentProps: ButtonPropsTable,
       },
     ]}
   />