+++ packages/terra-form-select/src/terra-dev-site/doc/form-select/Select.01.doc.jsx
@@ -3,12 +3,13 @@ import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/README.md';
 import { name, version } from '../../../../package.json';
 
-// Component Source
-import OptionSrc from '!raw-loader!../../../../src/shared/_Option';
-import OptGroupSrc from '!raw-loader!../../../../src/shared/_OptGroup';
-import SelectSrc from '!raw-loader!../../../../src/Select';
-
 // Example Files
+import SelectPropsTable from '!terra-props-table-loader!../../../../src/Select';
+import SelectSrc from '!raw-loader!../../../../src/Select';
+import OptionPropsTable from '!terra-props-table-loader!../../../../src/shared/_Option';
+import OptionSrc from '!raw-loader!../../../../src/shared/_Option';
+import OptGroupPropsTable from '!terra-props-table-loader!../../../../src/shared/_OptGroup';
+import OptGroupSrc from '!raw-loader!../../../../src/shared/_OptGroup';
 import AllowClearExample from '../example/legacy/AllowClear';
 import AllowClearExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/legacy/AllowClear';
 import ControlledExample from '../example/legacy/Controlled';
@@ -38,16 +39,23 @@ const Index = () => (
     packageVersion={version}
     readme={ReadMe}
     srcPath="https://github.com/cerner/terra-core/tree/master/packages/terra-form-select"
-    propsTables={[{
-      componentName: 'Select',
-      componentSrc: SelectSrc,
-    }, {
-      componentName: 'Option',
-      componentSrc: OptionSrc,
-    }, {
-      componentName: 'OptGroup',
-      componentSrc: OptGroupSrc,
-    }]}
+    propsTables={[
+      {
+        componentName: 'Select',
+        componentSrc: SelectSrc,
+        componentProps: SelectPropsTable,
+      },
+      {
+        componentName: 'Option',
+        componentSrc: OptionSrc,
+        componentProps: OptionPropsTable,
+      },
+      {
+        componentName: 'OptGroup',
+        componentSrc: OptGroupSrc,
+        componentProps: OptGroupPropsTable,
+      },
+    ]}
     examples={[{
       title: 'Default',
       description: 'A default variant allows selecting a single option.',