+++ packages/terra-form-textarea/src/terra-dev-site/doc/form-textarea/Textarea.1.doc.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/README.md';
+
 import { name, version } from '../../../../package.json';
 
-// Component Source
-import TextareaSrc from '!raw-loader!../../../../src/Textarea';
-
 // Example Files
+import FormTextareaPropsTable from '!terra-props-table-loader!../../../../src/Textarea';
+import FormTextareaSrc from '!raw-loader!../../../../src/Textarea';
 import ResizableExample from '../example/Resizable';
 import ResizableExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/Resizable';
 import FixedSmallSizeExample from '../example/FixedSmallSize';
@@ -32,8 +32,9 @@ const DocPage = () => (
     srcPath={`https://github.com/cerner/terra-core/tree/master/packages/${name}`}
     propsTables={[
       {
-        componentName: 'Textarea',
-        componentSrc: TextareaSrc,
+        componentName: 'Form Textarea',
+        componentSrc: FormTextareaSrc,
+        componentProps: FormTextareaPropsTable,
       },
     ]}
     examples={[