+++ packages/terra-form-textarea/src/terra-dev-site/doc/form-textarea/TextareaField.2.doc.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/TextareaField.md';
+
 import { name, version } from '../../../../package.json';
 
-// Component Source
-import TextareaFieldSrc from '!raw-loader!../../../../src/TextareaField';
-
 // Example Files
+import FormTextareaFieldPropsTable from '!terra-props-table-loader!../../../../src/TextareaField';
+import FormTextareaFieldSrc from '!raw-loader!../../../../src/TextareaField';
 import TextareaFieldExample from '../example/TextareaField';
 import TextareaFieldExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/TextareaField';
 import RequiredTextareaFieldExample from '../example/RequiredTextareaField';
@@ -20,8 +20,9 @@ const DocPage = () => (
     srcPath={`https://github.com/cerner/terra-core/tree/master/packages/${name}`}
     propsTables={[
       {
-        componentName: 'TextareaField',
-        componentSrc: TextareaFieldSrc,
+        componentName: 'Form Textarea Field',
+        componentSrc: FormTextareaFieldSrc,
+        componentProps: FormTextareaFieldPropsTable,
       },
     ]}
     examples={[