+++ packages/terra-profile-image/src/terra-dev-site/doc/profile-image/ProfileImage.1.doc.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
 import ReadMe from '../../../../docs/README.md';
+
 import { name, version } from '../../../../package.json';
 
-// Component Source
-import ImageSrc from '!raw-loader!../../../../src/ProfileImage';
-
 // Example Files
+import ProfileImagePropsTable from '!terra-props-table-loader!../../../../src/ProfileImage';
+import ProfileImageSrc from '!raw-loader!../../../../src/ProfileImage';
 import ProfileImageDefault from '../example/ProfileImageDefault';
 import ProfileImageDefaultSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/ProfileImageDefault';
 
@@ -26,7 +26,8 @@ const DocPage = () => (
     propsTables={[
       {
         componentName: 'Profile Image',
-        componentSrc: ImageSrc,
+        componentSrc: ProfileImageSrc,
+        componentProps: ProfileImagePropsTable,
       },
     ]}
   />